[Feat] - 결과/유니온 페이지 컴포넌트 분리
