+++ src/app/maple/result/union/page.tsx
@@ -1,37 +1,15 @@
 "use client";
 
 import React, { Suspense } from "react";
-import { useSearchParams } from "next/navigation";
 
-import UnionArtifact from "@/components/maple/result/union/UnionArtifact";
-import UnionRaider from "@/components/maple/result/union/UnionRaider";
+import UnionInfo from "@/components/maple/result/characterInfo/union/UnionInfo";
+
 import { Spinner } from "@nextui-org/spinner";
 
-import { useCharacterQuery } from "@/hooks/maple/useQuery";
-
 const UnionPage = () => {
-  const params = useSearchParams();
-
-  const name = params.get("character_name");
-
-  const { data } = useCharacterQuery(name, "union");
-
   return (
     <Suspense fallback={<Spinner />}>
-      <div className="flex flex-col gap-5 w-[1280px] md:w-full sm:w-full">
-        {data.unionArtifactInfo && (
-          <UnionArtifact
-            artifact={data.unionArtifactInfo}
-            level={data.unionCharacterInfo.union_artifact_level}
-          />
-        )}
-        {data.unionRaiderInfo && (
-          <UnionRaider
-            raider={data.unionRaiderInfo}
-            unionInfo={data.unionCharacterInfo}
-          />
-        )}
-      </div>
+      <UnionInfo />
     </Suspense>
   );
 };
